refactor(scheduler): extract review and meta-memory handlers (issue #224 task 3)
